# docs/openapi_client.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Caser Client API",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "description": "Create a bare global account. Used two ways: a lawyer/client\nregistering with no invite yet (an office manager attaches them to a\nfirm afterward), or — more commonly now — someone following an invite\nlink for an email with no Identity yet. In the second case, a\n*successful* registration for that exact email is itself the\nacceptance (CLAUDE.md's MembershipInvites note) — no separate\nconfirmation step, so every matching pending invite resolves to\naccepted immediately below.",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "lawyer/client login only — nobody cross-logs into the other app\n(CLAUDE.md's Roles section). An office_manager account exists too (same\nIdentities table), but its Membership.role at this tenant won't be\nLAWYER/CLIENT, so it's rejected here and pointed at the admin portal.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/lobby-login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Lobby Login",
        "description": "lawyer/client login from the lobby (www.<BASE_DOMAIN>) — CLAUDE.md's\nMulti-tenancy architecture. Mirrors admin_api's lobby-login: same\npassword check, but resolves active LAWYER/CLIENT memberships (at active\ntenants) instead of OFFICE_MANAGER ones. role is included per option\nsince it's needed to redirect straight into the right nav on the tenant\nsubdomain landing page — that page is a different origin, so it can't\nread anything stored here.",
        "operationId": "lobby_login_auth_lobby_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LobbyLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LobbyLoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/profile": {
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Update My Profile",
        "description": "Self-service only — bio/photo_url/years_of_experience are global to\nthe person, never a lever another party (an office_manager, CLAUDE.md's\nauthority boundary) gets to pull. Whether any of it is actually shown\npublicly is the separate per-membership show_on_public_page toggle.",
        "operationId": "update_my_profile_auth_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/change-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Change My Password",
        "description": "Self-service, for every role — see admin_api's identical route for\nthe full reasoning (shared/password_reset.py is the actual shared\nlogic; this is just this app's thin route on top of it).",
        "operationId": "change_my_password_auth_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Forgot Password",
        "description": "Always the same generic response regardless of whether the email\nmatched a real account (CLAUDE.md's PasswordResetTokens note — never\nreveal which emails are registered). See admin_api's identical route.",
        "operationId": "forgot_password_auth_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Reset Password",
        "operationId": "reset_password_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/dev-outbox": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Dev Outbox",
        "description": "Dev-only stand-in for real email delivery — see forgot_password above.",
        "operationId": "dev_outbox_auth_dev_outbox_get",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "properties": {
          "current_password": {
            "type": "string",
            "title": "Current Password"
          },
          "new_password": {
            "type": "string",
            "minLength": 8,
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "ChangePasswordRequest",
        "description": "Self-service \"change my password\" while logged in — every role can\ndo this for their own account. Never touches another person's account\n(see CLAUDE.md's office_manager authority boundary)."
      },
...
      "ForgotPasswordRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "ForgotPasswordRequest"
      },
...
      "GenericMessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "GenericMessageResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IdentityResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "years_of_experience": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Of Experience"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email"
        ],
        "title": "IdentityResponse"
      },
      "LobbyLoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LobbyLoginRequest",
        "description": "lawyer/client login from the lobby (www.<BASE_DOMAIN>) — no tenant\nsubdomain known yet, unlike LoginRequest's per-tenant version."
      },
      "LobbyLoginResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "tenants": {
            "items": {
              "$ref": "#/components/schemas/LobbyTenantOption"
            },
            "type": "array",
            "title": "Tenants"
          },
          "pending_invites": {
            "items": {
              "$ref": "#/components/schemas/PendingInviteOption"
            },
            "type": "array",
            "title": "Pending Invites",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "tenants"
        ],
        "title": "LobbyLoginResponse"
      },
      "LobbyTenantOption": {
        "properties": {
          "tenant_id": {
            "type": "integer",
            "title": "Tenant Id"
          },
          "subdomain": {
            "type": "string",
            "title": "Subdomain"
          },
          "firm_name": {
            "type": "string",
            "title": "Firm Name"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "subdomain",
          "firm_name",
          "role"
        ],
        "title": "LobbyTenantOption"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "PendingInviteOption": {
        "properties": {
          "invite_id": {
            "type": "integer",
            "title": "Invite Id"
          },
          "tenant_id": {
            "type": "integer",
            "title": "Tenant Id"
          },
          "subdomain": {
            "type": "string",
            "title": "Subdomain"
          },
          "firm_name": {
            "type": "string",
            "title": "Firm Name"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "type": "object",
        "required": [
          "invite_id",
          "tenant_id",
          "subdomain",
          "firm_name",
          "role"
        ],
        "title": "PendingInviteOption",
        "description": "A lawyer/client invite still awaiting this identity's response —\nsurfaced at lobby-login the same way active memberships are (CLAUDE.md's\nMembershipInvites note: \"the invite shows up as a pending action for\nthem the next time they log in\")."
      },
...
      "RegisterRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "description": "Create a bare global account — no firm attached yet. A lawyer/client\nuses this once, then an office manager attaches them to a firm via\nadmin_api's POST /members."
      },
      "ResetPasswordRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "new_password": {
            "type": "string",
            "minLength": 8,
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "token",
          "new_password"
        ],
        "title": "ResetPasswordRequest"
      },
      "SessionResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "role"
        ],
        "title": "SessionResponse"
      },
      "UpdateProfileRequest": {
        "properties": {
          "bio": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "years_of_experience": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 80.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Of Experience"
          }
        },
        "type": "object",
        "title": "UpdateProfileRequest",
        "description": "Self-service profile edit — bio/photo_url/years_of_experience are all\nself-reported and global to the person (CLAUDE.md's Identities note),\nnever a lever another party (e.g. an office_manager) gets to pull."
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "super_admin",
          "office_manager",
          "lawyer",
          "client"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}